Successful Insert-Into-Table tests
Refactored and simplified insertion of object values into an SQL table.
All insertion tests are running now. The code for creation of SQL tables and insertion of objects is complete (except for handling arrays, maps and sets).
Selection is still incomplete and untested.

The committed code is documented and can be used in the core project.