Fixed settings test.
Context:
Some simple property value tables do not require a key, e.g. the settings property table.
Therefore, in Java, we introduced the NoneConverter. However, when the property table is
created, the database conversion mechanism notices that the subject, to which the property belongs,
must potentially be represented in the table. This should only be true for subjects that have fields that
are not of the type 'None' (-> if the property table potentially contains more than one field and therefore
requires primary keys).

The fix:
When finding the table constraints, we check whether a field of the converter is of type 'Subject'.
If the type, however, does not contain any convertable fields (e.g., only fields that are of type None),
we do not create any table constraints.